[#4400] period locking all done
